Update ch5.md
I have read this document various times. Using "a" as the name of the new instance of object "Foo" is very confusing. The new object's name can be "Foo2" or other name related to "Foo" function. So I suggest using "Foo2". I believe every other names is better that "a" since sometimes it is very confusing and reader has to return and read the previous paragraph and code peaces again and again.