fixed logical mistake in explanation of iterator loop
The description states, that res.done is consulted in each iteration of the `for equivalent of a for..of loop` example and if res.done is `false`, the iteration doesn't occur. In fact, it's the other way around: if res.done is `true`, the statement evaluates to false and the iteration doesn't occur.