Typos with effect / affecting
I notice you refer to operators with side effects a lot, and use the verb `side-effecting` while I can understand that it would be logical to simply add `ing` to the end of `side-effect`, that verb doesn't actually exist. Technically, you should be using `affecting` instead. However, one could make the case that using the term "side-effecting" makes it clearer for the readers to understand that it is simply a reference to operators with side-effects, which are defined beforehand. It's up to you to decide, though I submitted the orthographically correct changes in this commit in case you want to correct it.