Difference in Node environment
Not sure if the implicit understanding is that we should follow along using the js environment on our browsers, but if not it would be good to note that console.log ( this.a ) would return undefined in the node environment. May not be obvious to newcomers and who knows maybe even 'seasoned' js devs ( Kanye Shrug ).