Update ch5.md
I've had no problem following the book, but on line 472 there is a syntax construction that has me completely baffled:

    deps[i] = modules[deps[i]];

I checked and this works, but I can't figure out exactly what it does or even why it's legal. It seems to be referencing a non-existent modules property, which should assign undefined to that slot in deps. I don't see how making all items in deps undefined does anything useful.  

Could you put in a word about what it's doing and how, please? :-)