Update ch5.md
I think using the get() function in the define() function is better. It makes us easier to understand that we want to set dependent modules to the deps array. It helps us to have one place to change if you want to check if the dependant module is already available. Also it shows us that we can use the get() function inside the define() function to consolidate what we are learning about closure scope. Thanks.