Swap 'data accessor' with 'data descriptor'
I think that 'data descriptor' may be a more accurate description of the property than 'data accessor'. A refresher that 'accessor descriptors' are the properties with getter and setter functions may also help solidify the difference between 'data descriptors' and 'accessor descriptors' for readers who are new to these concepts. Personally, as someone new to the concept of property descriptors, I had trouble keeping track of the meanings of each of the terms presented in chapter 3: 'data accessor', 'accessor descriptor', 'accessor properties', 'data descriptor', 'property descriptor'.