ch3: clarify what's being iterated over in a classic for loop
The previous statement was that the classic 'for' loop iterated over the array indices, but that's just coincidental due to setting the limit of i from 0 to the array's length. There's nothing special about that choice, as it could easily be written so that i goes from 0 to 1, or -5 to +5. Therefore, just added a slight clarification so that there's no implication that a regular for loop has anything particularly to do with iterating over an array's indices, just that it's often been used for such a purpose.