simple mistake in code corrected
below the "Function Expressions Immediately" section, the text talks about two forms of immediately invoking a function expression. but both the codes are using the traditional form. (they both place the invoke operator() outside of the wrapping parentheses.)
In the second snippet I moved the parentheses of invoking, into the wrapping parentheses.