Update chapter 2
In the declaration of the class "Book" you declared the constructor to accept an Object with 3 properties (title, author, publishedOn), but then when you instanciate that class, you use the 3 values not wrapped into an object. Since in all the rest of the examples you did not use wrapping objects, I think it's best to correct the class declaration instead of fixing the "new Book(...)"  statement.