Replace the "As mentioned earlier" part in ch1.md
Maybe I'm wrong (hope I'm not), I don't remember any mentions about an
array type at that point of the book. I've tried to find it, but I found
nothing. It might be an artefact of the original book (the 1st edition).

So I'd replace the "As mentioned earlier" part with something not
confusing. I'm not a native English, so probably you have a better idea
for the replacement.